#ROBOMERGE-OWNER: robert.manuszewski
#ROBOMERGE-AUTHOR: bob.tellez
#ROBOMERGE-SOURCE: CL 12146920 via CL 12149228 via CL 12149307
#ROBOMERGE-BOT: CORE (Main -> Dev-Core) (v662-12191386)

[CL 12194366 by bob tellez in Dev-Core branch]

+++ Engine/Source/Runtime/Core/Private/HAL/MallocBinned2.cpp
@@ -53,7 +53,7 @@ static FAutoConsoleVariableRef GMallocBinned2AllocExtraCVar(
 
 #endif
 
-double GMallocBinned2FlushThreadCacheMaxWaitTime = 0.02f;
+float GMallocBinned2FlushThreadCacheMaxWaitTime = 0.02f;
 static FAutoConsoleVariableRef GMallocBinned2FlushThreadCacheMaxWaitTimeCVar(
 	TEXT("MallocBinned2.FlushThreadCacheMaxWaitTime"),
 	GMallocBinned2FlushThreadCacheMaxWaitTime,
